refactor for common code
